Move Markdown format to its own, new project.
The new project is webapi-markdown. It focuses on the description of
the markdown format, as well as being a home for .md files for various
services.

That new project is not specific to Racket. Conceivably people could
submit pull requests for various web services, and use those files wih
parsers and web FFIs written in languages other than Racket.

This original project _is_ specific to Racket. Although it still uses
those markdown service files, the proper home for those files is their
own project.