`wffi-define-all`: handle lib name simply. Closes #2.
Automatically treat the `lib-name` as relative to the path of the
source file using the `wffi-define-all` form. For example in
`my-api.rkt` you can simply write:

    (wffi-define-all "my-api.md" <before> <after>)

and if `my-api.md` is in the same directory as `my-api.rkt` it will
just work.

The contortions @takikawa had to use (for both phases 0 and 1) are no
longer required. (To avoid breaking such code, the
`current-markdown-files-path` parameter is still `provide`d, but it's
a deprecated no-op.)