Support repeating parameters.
Some web APIs need this. For example a `foo` query parameter may be
supplied multiple times, with different values.

Since we're using a `dict`, we represent this as still just one dict
entry, but if the value is list? then we spread it out into multiple
query parameters. Note this is also supported for headers, which is
good because some HTTP headers may be repeated (e.g. cookies).